cdk/requirements.txt
aws-cdk.aws-certificatemanager
aws-cdk.aws-cloudformation
aws-cdk.aws-cloudfront
aws-cdk.aws-cloudwatch
aws-cdk.aws-cognito
aws-cdk.aws-ec2
aws-cdk.aws-elasticbeanstalk
aws-cdk.aws-events
aws-cdk.aws-iam
aws-cdk.aws-kms
aws-cdk.aws-lambda
aws-cdk.aws-logs
aws-cdk.aws-rds
aws-cdk.aws-route53
aws-cdk.aws-route53-targets
aws-cdk.aws-s3
aws_cdk.aws-s3-notifications
aws-cdk.aws-sam
aws-cdk.aws-secretsmanager
aws-cdk.aws-sns
aws-cdk.aws-sns-subscriptions
aws-cdk.aws-sqs
aws-cdk.aws-ssm
aws-cdk.core
aws-cdk.cx-api
aws-cdk.region-info
boto3                  
botocore              
PyYAML         
pytest
python-dateutil 
